Fix IE9 throwing a servlet exception on using special characters in Textfields
Modified patch from Andre published at http://echo.nextapp.com/site/node/6658
"I'm using the Extras.RichTextArea in my application. It works in most browsers. In IE9 I
got an exception when the input of the textarea is parsed on server side. The input source
tried to read the content as UTF-8. The internet explorer here created the editable iframe
with a document that has CP-1252 as charset encoding. Using characters like "äüö"
resulted in a crash of the application with the next try to sync.
For IE it is necessary to set the charset to UTF-8 after creating the document of the iframe."