Echo configuration via system & servlet properties
This change introduces a ServerConfiguration class that provides
access to both system and servlet init parameters, which can be used
to configure certain Echo behaviour. Existing code that read only system
properties was migrated to use the new class instead.

If a system property is specified, it takes precedence over a servlet
init parameter with the same name, which means this change should be
backwards compatible with existing Echo applications.

Additionally, the IE_EDGE_MODE configuration option was added. If enabled,
Echo will include a X-UA-Compatible meta tag with value 'IE=edge', which
tells Internet Explorer to always use the latest rendering engine.
The default setting is enabled.

Resolves #63
Resolved #56