Configurable handling of invalid Ajax responses
This feature introduces configurable handling of invalid Ajax responses returned
to the remote Echo client. This can happen if, for example, a web access management
agent intercepts the Echo synchronization Ajax request and returns a HTML login page
instead. This page would be displayed as HTML code in the Echo error window. With this patch,
the application can be configured to either restart immediately (thus forcing a GET request
that should lead to the login page) or to forward the browser to an arbitrary URI.