Fix overriden scrollpositions on focus components out of viewport.
This is a semi-dirty hack: Echo sets the focused component using
a timed process *after* all rendering occured (probably to circumvent
browser bugs.

But on changing focus the browsers scroll so the focus gets visible.
So we need to scroll back, afterwards.

*Update:* Turns out this workaround won't work for alle cases. A
implementation similar to the focus management is needed. So this is
only an iterims solution.